requirements.txt
# ───────────────────────────────────────────
# 핵심 의존성 (LLM & RAG)
# ───────────────────────────────────────────
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1

# ───────────────────────────────────────────
# 트랜스포머 임베딩 (발표자료 권장: BAAI/bge-m3)
# 최초 설치 후 모델 자동 다운로드 (~2GB)
# ───────────────────────────────────────────
sentence-transformers>=2.7.0

# ───────────────────────────────────────────
# 벡터 데이터베이스
# chromadb: Dense 임베딩 저장 및 유사도 검색
# ───────────────────────────────────────────
chromadb>=0.5.0

# ───────────────────────────────────────────
# PDF 처리
# pymupdf(fitz): 텍스트 레이어 추출 + 이미지 렌더링
# pdfplumber:    표(Table) 구조 추출
# ───────────────────────────────────────────
pymupdf>=1.23.0
pdfplumber>=0.11.0

# ───────────────────────────────────────────
# OCR (스캔 PDF 텍스트 추출)
# pytesseract: Tesseract OCR Python 래퍼
# Pillow:      이미지 처리 (OCR 입력용)
# ───────────────────────────────────────────
pytesseract>=0.3.10
Pillow>=10.0.0

# ───────────────────────────────────────────
# UI
# ───────────────────────────────────────────
streamlit>=1.30.0

# ───────────────────────────────────────────
# 유틸리티
# python-dotenv: .env 파일 로드
# google-generativeai: Gemini API
# ───────────────────────────────────────────
python-dotenv>=1.0.0
google-generativeai>=0.8.0
